requirements.txt
celery==5.5.3
Flask==2.2.5
gunicorn==23.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4
pyahocorasick==2.1.0
pymysql==1.1.1
pytz==2025.2
prometheus-client==0.21.0
redis==4.3.4
web3==7.12.0
cryptography==45.0.5